更新历史搜索记录

+++ app/src/main/java/com/wudaokou/easylearn/data/SearchRecord.java
@@ -6,13 +6,13 @@ import androidx.room.PrimaryKey;
 @Entity(tableName = "search_record")
 public class SearchRecord {
     @PrimaryKey
-    public int timestamp;
+    public long timestamp;
 
     public String content;
 
     public String subject;
 
-    public SearchRecord(final int timestamp, final String content,
+    public SearchRecord(final long timestamp, final String content,
                          final String subject) {
         this.timestamp = timestamp;
         this.content = content;
